Adds 1.2 paths and upgrade reasoner pydantic models
