+++ PLATER/examples/reasoner-trapi-1.2.json
@@ -1,0 +1,31 @@
+{
+   "message": {
+      "query_graph": {
+         "nodes": {
+            "n0": {
+               "categories": [
+                  "biolink:Gene"
+               ]
+            },
+            "n1": {
+               "ids": [
+                  "MONDO:0004969"
+               ],
+               "categories": [
+                  "biolink:Disease"
+               ]
+            }
+         },
+         "edges": {
+            "e0": {
+               "subject": "n0",
+               "object": "n1",
+               "predicates": [
+                  "biolink:related_to"
+               ]
+            }
+         }
+      }
+   },
+   "workflow": [{"id":  "lookup"}]
+}
