+++ PLATER/requirements.txt
@@ -7,7 +7,7 @@ starlette==0.14.2
 uvicorn==0.14.0
 reasoner-transpiler==1.8.1
 bmt-lite-2.1.0==2.1.1
-reasoner-pydantic==1.1.2.1
+reasoner-pydantic==1.1.2.2
 httpx==0.18.2
 pytest-httpx==0.12.0
 jsonasobj==1.3.1
