+++ PLATER/services/app_trapi_1_2.py
@@ -7,8 +7,8 @@ from PLATER.services.util.graph_adapter import GraphInterface
 from PLATER.services.util.question import Question
 from PLATER.services.util.api_utils import get_graph_interface, construct_open_api_schema, get_example
 
-# Mount open api at /1.1/openapi.json
-APP_TRAPI_1_1 = FastAPI(openapi_url="/openapi.json", docs_url="/docs", root_path='/1.1')
+# Mount open api at /1.2/openapi.json
+APP_TRAPI_1_2 = FastAPI(openapi_url="/openapi.json", docs_url="/docs", root_path='/1.2')
 
 
 async def get_meta_knowledge_graph(
@@ -22,19 +22,21 @@ async def get_meta_knowledge_graph(
 async def reasoner_api(
         request: ReasonerRequest = Body(
             ...,
-            example={"message": get_example("reasoner-trapi-1.1")},
+            example=get_example("reasoner-trapi-1.2"),
         ),
         graph_interface: GraphInterface = Depends(get_graph_interface),
 ):
     """Handle TRAPI request."""
     request_json = request.dict(by_alias=True)
+    workflow = request_json.get('workflow', [{"id": "lookup"}])
+    # @TODO handle workflow
     question = Question(request_json["message"])
     response = await question.answer(graph_interface)
-    request_json.update({'message': response})
+    request_json.update({'message': response, 'workflow': workflow})
     return request_json
 
 
-APP_TRAPI_1_1.add_api_route(
+APP_TRAPI_1_2.add_api_route(
     "/meta_knowledge_graph",
     get_meta_knowledge_graph,
     methods=["GET"],
@@ -44,7 +46,7 @@ APP_TRAPI_1_1.add_api_route(
     tags=["trapi"]
 )
 
-APP_TRAPI_1_1.add_api_route(
+APP_TRAPI_1_2.add_api_route(
     "/query",
     reasoner_api,
     methods=["POST"],
@@ -54,4 +56,4 @@ APP_TRAPI_1_1.add_api_route(
     tags=["trapi"]
 )
 
-APP_TRAPI_1_1.openapi_schema = construct_open_api_schema(app=APP_TRAPI_1_1, trapi_version="1.1.1", prefix='/1.1')
+APP_TRAPI_1_2.openapi_schema = construct_open_api_schema(app=APP_TRAPI_1_2, trapi_version="1.2.0", prefix='/1.2')
