+++ PLATER/services/server.py
@@ -2,12 +2,12 @@
 import os
 
 from fastapi import  FastAPI
-from fastapi.openapi.utils import get_openapi
 from starlette.middleware.cors import CORSMiddleware
 from PLATER.services.config import config
 from PLATER.services.util.logutil import LoggingUtil
-from PLATER.services.app_trapi_1_0 import APP_TRAPI_1_0
+from PLATER.services.app_common import APP_COMMON
 from PLATER.services.app_trapi_1_1 import APP_TRAPI_1_1
+from PLATER.services.app_trapi_1_2 import APP_TRAPI_1_2
 from PLATER.services.util.api_utils import construct_open_api_schema
 
 TITLE = config.get('PLATER_TITLE', 'Plater API')
@@ -23,14 +23,18 @@ APP = FastAPI()
 
 # Mount 1.1 app at /1.1
 APP.mount('/1.1',  APP_TRAPI_1_1, 'Trapi 1.1')
-# Mount default  1.0 app at /
-APP.mount('/', APP_TRAPI_1_0, 'Trapi 1.0')
+# Mount 1.2 app at /1.2
+APP.mount('/1.2',  APP_TRAPI_1_2, 'Trapi 1.2')
+# Mount default app at /
+APP.mount('/', APP_COMMON, '')
 # Add all routes of each app for open api generation at /openapi.json
 # This will create an aggregate openapi spec.
 APP.include_router(APP_TRAPI_1_1.router, prefix='/1.1')
-APP.include_router(APP_TRAPI_1_0.router)
-# Construct app /openapi.json
-APP.openapi_schema = construct_open_api_schema(app=APP, trapi_version='1.0')
+APP.include_router(APP_TRAPI_1_2.router, prefix='/1.2')
+APP.include_router(APP_COMMON.router)
+# Construct app /openapi.json # Note this is not to be registered on smart api . Instead /1.1/openapi.json
+# or /1.2/openapi.json should be used.
+APP.openapi_schema = construct_open_api_schema(app=APP, trapi_version='N/A')
 
 # CORS
 APP.add_middleware(
